requirements.txt
# Core Dependencies
fastapi==0.95.2
uvicorn==0.22.0

# Database
sqlalchemy==1.4.46
databases==0.7.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.3.1
httpx==0.24.1

# Optional (for async database support)
asyncpg==0.27.0